+++ RELEASE_NOTES.md
@@ -1,5 +1,6 @@
 #### 2.1.2 - Unreleased
 * Added constants for more HTTP methods.
+* Added fix for `thead` element without nested `tr` element
 
 #### 2.1.1 - December 24 2014
 * Add SkipRows parameter to CsvProvider.